+++ crates/store/src/lib.rs
@@ -1,10 +1,9 @@
-use std::time::Duration;
-
 mod accounts;
 mod blocks;
 mod db;
 mod errors;
 pub mod genesis;
+mod inner_forest;
 mod server;
 pub mod state;
 
@@ -15,6 +14,3 @@ pub use server::{DataDirectory, Store};
 // CONSTANTS
 // =================================================================================================
 const COMPONENT: &str = "miden-store";
-
-/// How often to run the database maintenance routine.
-const DATABASE_MAINTENANCE_INTERVAL: Duration = Duration::from_secs(24 * 60 * 60);